requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-cors-headers==3.10.1
django-crontab==0.7.1
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
pycparser==2.21
pycryptodome==3.14.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.26.9
wakeonlan==2.1.0
xlrd==2.0.1
xlwt==1.3.0
